ci: refactor 'release_artifacts' handling into a separate transform
We'll want other tasks to make use of this 'release-artifacts' feature
soon as well, so move them outside of the build transforms and into a
new transform module.
